pyproject.toml
[project]
name = "mro_agenticops"
version = "0.1.0"
description = "AgenticOps LLMOps + RAG local with Ollama"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name" }]
keywords = ["llmops", "rag", "agentic", "mlops"]

dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "mlflow>=2.8",
  "pytest>=7.0",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "ollama>=0.2.0",
  "requests>=2.31",
  "sentence-transformers>=2.3",
  "faiss-cpu>=1.7.4",
  "pypdf>=3.17",
  "pypdfium2>=4.26"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
